size_t

+++ libft/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: akroll <akroll@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 10:50:31 by akroll            #+#    #+#             */
-/*   Updated: 2022/03/24 13:17:42 by akroll           ###   ########.fr       */
+/*   Updated: 2022/03/28 10:35:17 by akroll           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 //locate byte c in byte string s within the first n bytes
 
-void *ft_memchr(const void *s, int c, unsigned int n)
+void *ft_memchr(const void *s, int c, size_t n)
 {
 	const unsigned char *p;
