+++ libft/ft_memcpy.c
@@ -6,15 +6,17 @@
 /*   By: akroll <akroll@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/23 12:34:16 by akroll            #+#    #+#             */
-/*   Updated: 2022/03/25 13:28:25 by akroll           ###   ########.fr       */
+/*   Updated: 2022/03/28 10:17:10 by akroll           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void *ft_memcpy(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void *ft_memcpy(void *dest, const void *src, size_t n)
 {
 	const unsigned char *p_src;
 	unsigned char *p_dest;
-	unsigned int i;
+	size_t i;
 
 	p_dest = dest;
 	p_src = src;
